requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
cached-property==1.5.1
certifi==2020.4.5.1
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
defusedxml==0.6.0
django-allauth==0.42.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-filter==2.3.0
django-pragmatic==3.2.2
django-range-slider==0.2.4
django-simple-history==2.11.0
djangorestframework==3.11.1
easy-thumbnails==2.7
idna==2.9
isodate==0.6.0
lxml==4.5.2
Markdown==3.2.2
oauthlib==3.1.0
Pillow==7.1.2
py-wsse==0.1
pycparser==2.20
pyOpenSSL==19.1.0
python-decouple==3.3
python-pragmatic==0.4.2
python-tbk==0.2.2
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.3.1
stripe==2.48.0
urllib3==1.25.9
zeep==3.4.0
